Move EntityConstructor type to entities.ts
